app/requirements.txt
agno>=0.1.0
agno-cli>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.3
mistralai>=1.9.10
# Para base de conhecimento
sqlalchemy>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Para webhook server
flask>=2.3.0
requests>=2.31.0
# Para processar PDFs com imagens
pypdf>=3.0.0
pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

